Update patch set 1
Patch Set 1:

That doesn't seem to be exactly the case.
have you tried to boot p3100 with
androidboot.carrier=wifi? which is the extra flag
the p3110 uses.

It overrides the bootloader and the modem won't start.
You shouldn't do that because, you probably would
want your modem to start, lol, but never the less.

Patch-set: 1