Update patch set 2
Patch Set 2: Code-Review+1

Yeah this is what I use on my builds and IR even works,  daavvis just mentioned something codeworkx and Andrew, and
I don't know how to upload a patch that is technically a new project, but shouldn't we add consumerir to /external?

Right now every device that uses the new consumerir has the
same directory 'consumerir' redundantly in the device repo.

Would have to add /external/consumerir then change the device
repos, but as it is, this is correct.

Patch-set: 2
Label: Code-Review=+1